Update 1.0.3: Bugs fixed and some changes.
The show message function has been removed for the moment. Because it caused the program died. Why? I think it's because when the program is running as a daemon there isn't a window parent to be passed to initialize the dialog and then that function cause the program die. It's strange, because that function's documentation says it allows a null parent, but anyway the program dies when it's passed. On other hands, a bug has been fixed when the message signal is sent to the bus. It was sent two times and it was reflected when a suspend signal is sent. It caused the system be suspended two times. By last, if the GUI is used, it will charge the icons from the current directory first, and if they aren't found there, then the current directory will be changed to /usr/share/spm, the which is program's resources directory. It's so, when the program is installed, obviously!