1.0.8: Added new feature.
It notifies you when it's monitoring your battery charge and it's
less than 25%, and also when it will put your computer to sleep.
i.e. when your laptop's battery charge has less than 15%

Also, the README was modified improving the required dependencies.

By last but not least, every header file was modified. All the
compile-time verification were removed because, as far as I know,
the compilers already do that, so I think it's unnecessary.