1.0.9: Some features were added.
Added a new function for storing formatted strings, so avoid
the hard-coded way of doing it used before.

Also a lot of changes in the battery_monitor's source file were made
for improving its legibility, or at least trying to do it.

By last, but not least in the GUI's buttons if you put your mouse cursor
on them, a short message will be shown.